+++ kaithem/var/modules/data/Examples/Direct Cherrypy Page.py
@@ -2,7 +2,7 @@
 ## If manually editing, you must reload the code. Delete the resource timestamp so kaithem knows it's new
 __data__="""
 {continual: false, enable: true, once: true, priority: interactive, rate-limit: 0.0,
-  resource-timestamp: 1568112911280269, resource-type: event}
+  resource-timestamp: 1591304183645281, resource-type: event}
 
 """
 
@@ -11,7 +11,7 @@ __trigger__='False'
 if __name__=='__setup__':
     #This code runs once when the event loads. It also runs when you save the event during the test compile
     #and may run multiple times when kaithem boots due to dependancy resolution
-    __doc__='Create a page directly in code as you would in a traditional cherrypy app. Go to /hello to test this.'
+    __doc__='Create a page directly in code as you would in a traditional cherrypy app. Go to /hello_example/ to test this.'
     
     
     import cherrypy
@@ -22,7 +22,7 @@ if __name__=='__setup__':
             return "Hello World!"
     
     hello = HelloWorld()
-    kaithem.web.controllers[None]= hello
+    kaithem.web.controllers[("hello_example",)]= hello
 
 def eventAction():
     pass